fix recursive dft formatting & remove duplicate file
